Use ResourceIamPolicy for google_project_iam_policy (#3410)
