+++ google-beta/iam_project.go
@@ -18,6 +18,17 @@ var IamProjectSchema = map[string]*schema.Schema{
 	},
 }
 
+// In google_project_iam_policy, project is required and not inferred by
+// getProject.
+var IamPolicyProjectSchema = map[string]*schema.Schema{
+	"project": {
+		Type:             schema.TypeString,
+		Required:         true,
+		ForceNew:         true,
+		DiffSuppressFunc: compareProjectName,
+	},
+}
+
 type ProjectIamUpdater struct {
 	resourceId string
 	Config     *Config
@@ -33,6 +44,15 @@ func NewProjectIamUpdater(d *schema.ResourceData, config *Config) (ResourceIamUp
 
 	return &ProjectIamUpdater{
 		resourceId: pid,
+		Config:     config,
+	}, nil
+}
+
+// NewProjectIamPolicyUpdater is similar to NewProjectIamUpdater, except that it
+// doesn't call getProject and only uses an explicitly set project.
+func NewProjectIamPolicyUpdater(d *schema.ResourceData, config *Config) (ResourceIamUpdater, error) {
+	return &ProjectIamUpdater{
+		resourceId: d.Get("project").(string),
 		Config:     config,
 	}, nil
 }
@@ -83,4 +103,13 @@ func (u *ProjectIamUpdater) GetMutexKey() string {
 
 func (u *ProjectIamUpdater) DescribeResource() string {
 	return fmt.Sprintf("project %q", u.resourceId)
+}
+
+func compareProjectName(_, old, new string, _ *schema.ResourceData) bool {
+	// We can either get "projects/project-id" or "project-id", so strip any prefixes
+	return GetResourceNameFromSelfLink(old) == GetResourceNameFromSelfLink(new)
+}
+
+func getProjectIamPolicyMutexKey(pid string) string {
+	return fmt.Sprintf("iam-project-%s", pid)
 }
